requirements.txt
gunicorn
psycopg2
django
django-cors-headers
djangorestframework
django-filter
django-restql
drf-guard